rename(RacingCarPlayer) : move RacingCarPlayer to Controller package
